Crate rpf

rpf is a framework to help create Rust programs

Modules

as_path

pathmod

prog

status

styled

test

Structs

Prog
A structure for storing program information

Enums

Color
Wrapper for ansi_term::Colour

ExitStatus
Enum used for setting exit statuses

Traits

AsPath
A trait for treating String and str as Path and PathBuf

Exit
Causes a Prog struct to exit using an error and exit status

PathMod
Adds some useful functions for manipulating and retrieving information from paths

Styled
Trait for creating stylized console printing using ansi_term