Struct rpf::prog::Prog

pub struct Prog {
    pub name: &'static str,
    pub vers: &'static str,
    pub yr: &'static str,
}

A structure for storing program information

Example

use rpf::Prog;

static UTIL: Prog = Prog { name: "util", vers: "0.1.0", yr: "2015" };
assert_eq!(UTIL.name, "util");
assert_eq!(UTIL.vers, "0.1.0");
assert_eq!(UTIL.yr, "2015");

Fields

name

Name of program

vers

Version of program

yr

Year of copyright for program

Methods

impl Prog

fn copyright(&self, license: &str, auth: &[&str])

Prints copyright, version, and author information

Example

use rpf::Prog;

let prog = Prog { name: "util", vers: "0.1.0", yr: "2015" };
prog.copyright(
"Copyright (C) 2015 util developers\n\
License GPLv3+: GNU GPL version 3 or later <http://gnu.org/licenses/gpl.html>.\n\
This is free software: you are free to change and redistribute it.\n\
There is NO WARRANTY, to the extent permitted by law.\n", &["Author"]);

Would print the following:

util 0.1.0
Copyright (C) 2015 util developers
License GPLv3+: GNU GPL version 3 or later <http://gnu.org/licenses/gpl.html>.
This is free software: you are free to change and redistribute it.
There is NO WARRANTY, to the extent permitted by law.

Written by Author

fn prog_try(&self)

Used for when no arguments are given to a util

Example

use rpf::Prog;

let prog = Prog { name: "util", vers: "0.1.0", yr: "2015" };
prog.prog_try();

Would print the following:

util: Missing arguments
Try util --help for more information

fn exit(&self, status: ExitStatus)

Wrapper for process::exit, immediately exits the process with the set exit status.

Example

use rpf::{Prog,Exit,ExitStatus};

let prog = Prog { name: "test", vers: "0.1.0", yr: "2015" };
prog.exit(ExitStatus::Ok);