Trait rpf::status::Exit

pub trait Exit<T: AsRef<str>> {
    fn error(&self, mesg: T, status: ExitStatus);
    fn path_error(&self, mesg: T, item: PathBuf);
}

Causes a Prog struct to exit using an error and exit status

Required Methods

fn error(&self, mesg: T, status: ExitStatus)

Used for errors, prints error messages in red terminal font and calls rpf::Exit::exit

Example

use rpf::{Prog,Exit,ExitStatus};

let prog = Prog { name: "test", vers: "0.1.0", yr: "2015" };
prog.error("Some kind of error occured!".to_string(), ExitStatus::Error);

fn path_error(&self, mesg: T, item: PathBuf)

Used for errors when working with paths, works similar to error